Refactor networking things
